openpilot v0.9.7 release
date: 2024-04-07T10:20:32
master commit: c95404d564a9a8d168f8f587887e9c33f96db861

+++ panda/board/stm32h7/lli2c.h
@@ -16,10 +16,10 @@ bool i2c_write_reg(I2C_TypeDef *I2C, uint8_t addr, uint8_t reg, uint8_t value) {
   bool ret = false;
   for(uint32_t i=0U; i<10U; i++) {
     register_clear_bits(&I2C->CR2, I2C_CR2_ADD10);
-    I2C->CR2 = ((addr << 1U) & I2C_CR2_SADD_Msk);
+    I2C->CR2 = ((uint32_t)addr << 1U) & I2C_CR2_SADD_Msk;
     register_clear_bits(&I2C->CR2, I2C_CR2_RD_WRN);
     register_set_bits(&I2C->CR2, I2C_CR2_AUTOEND);
-    I2C->CR2 |= (2 << I2C_CR2_NBYTES_Pos);
+    I2C->CR2 |= 2UL << I2C_CR2_NBYTES_Pos;
 
     I2C->CR2 |= I2C_CR2_START;
     if(!i2c_status_wait(&I2C->CR2, I2C_CR2_START, 0U)) {
@@ -61,10 +61,10 @@ bool i2c_read_reg(I2C_TypeDef *I2C, uint8_t addr, uint8_t reg, uint8_t *value) {
   bool ret = false;
   for(uint32_t i=0U; i<10U; i++) {
     register_clear_bits(&I2C->CR2, I2C_CR2_ADD10);
-    I2C->CR2 = ((addr << 1U) & I2C_CR2_SADD_Msk);
+    I2C->CR2 = ((uint32_t)addr << 1U) & I2C_CR2_SADD_Msk;
     register_clear_bits(&I2C->CR2, I2C_CR2_RD_WRN);
     register_clear_bits(&I2C->CR2, I2C_CR2_AUTOEND);
-    I2C->CR2 |= (1 << I2C_CR2_NBYTES_Pos);
+    I2C->CR2 |= 1UL << I2C_CR2_NBYTES_Pos;
 
     I2C->CR2 |= I2C_CR2_START;
     if(!i2c_status_wait(&I2C->CR2, I2C_CR2_START, 0U)) {
@@ -92,7 +92,7 @@ bool i2c_read_reg(I2C_TypeDef *I2C, uint8_t addr, uint8_t reg, uint8_t *value) {
   I2C->TXDR = reg;
 
   // Restart
-  I2C->CR2 = (((addr << 1) | 0x1U) & I2C_CR2_SADD_Msk) | (1U << I2C_CR2_NBYTES_Pos) | I2C_CR2_RD_WRN | I2C_CR2_START;
+  I2C->CR2 = (((addr << 1) | 0x1U) & I2C_CR2_SADD_Msk) | (1UL << I2C_CR2_NBYTES_Pos) | I2C_CR2_RD_WRN | I2C_CR2_START;
   ret = i2c_status_wait(&I2C->CR2, I2C_CR2_START, 0U);
   if(!ret) {
     goto end;
